ai-services/requirements.txt
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.2
kafka-python==2.0.2
asyncio==3.4.3
aiohttp==3.9.1
